lstm with rouped normalization layers

+++ llab/cpu/model.c
@@ -2270,7 +2270,7 @@ void update_model(model* m, float lr, float momentum, int mini_batch_size, int g
     if(m == NULL)
         return;
     
-    lambda*=mini_batch_size;
+    lambda*=(float)mini_batch_size;
     
     if(regularization == L2_REGULARIZATION){
         add_l2_residual_layer(m,total_number_weights,lambda);
